api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
PyYAML==6.0.1
python-multipart==0.0.6
